import-excel

+++ server/database/migrations/000007_create_books_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->longText('description')->nullable();
             $table->Integer('unit_price')->nullable();
             $table->bigInteger('quantity')->nullable();
-            $table->double('weight')->nullable();
+            $table->Integer('weight')->nullable();
             $table->string('format')->nullable();
             $table->year('year')->nullable();
             $table->string('language')->nullable();
